Trait chart::Backend

pub trait Backend {
    type Error;

    const EXT: &'static str = "";

    fn header(&mut self, max_x: i32, max_y: i32) -> Result<(), Self::Error>;
    fn line(
        &mut self, 
        start_x: i32, 
        end_x: i32, 
        start_y: i32, 
        end_y: i32, 
        style: &str
    ) -> Result<(), Self::Error>;
    fn text(
        &mut self, 
        x: i32, 
        y: i32, 
        content: impl Display, 
        orientation: Orientation
    ) -> Result<(), Self::Error>;
    fn node(&mut self, x: i32, y: i32, n: usize) -> Result<(), Self::Error>;
    fn structline(
        &mut self, 
        source: (i32, i32, usize), 
        target: (i32, i32, usize), 
        style: Option<&str>
    ) -> Result<(), Self::Error>;

    fn init(&mut self, max_x: i32, max_y: i32) -> Result<(), Self::Error> { ... }
    fn structline_matrix(
        &mut self, 
        source: (i32, i32), 
        target: (i32, i32), 
        matrix: Vec<Vec<u32>>, 
        class: Option<&str>
    ) -> Result<(), Self::Error> { ... }
}

Required Associated Types

type Error

Provided Associated Constants

const EXT: &'static str = ""

If the backend writes to a file, this is the extension commonly taken by the file type

Required Methods

fn header(&mut self, max_x: i32, max_y: i32) -> Result<(), Self::Error>

fn line(
&mut self,
start_x: i32,
end_x: i32,
start_y: i32,
end_y: i32,
style: &str
) -> Result<(), Self::Error>

fn text(
&mut self,
x: i32,
y: i32,
content: impl Display,
orientation: Orientation
) -> Result<(), Self::Error>

fn node(&mut self, x: i32, y: i32, n: usize) -> Result<(), Self::Error>

fn structline(
&mut self,
source: (i32, i32, usize),
target: (i32, i32, usize),
style: Option<&str>
) -> Result<(), Self::Error>

Provided Methods

fn init(&mut self, max_x: i32, max_y: i32) -> Result<(), Self::Error>

fn structline_matrix(
&mut self,
source: (i32, i32),
target: (i32, i32),
matrix: Vec<Vec<u32>>,
class: Option<&str>
) -> Result<(), Self::Error>